apply() and call() method🍔
The main question here is: how do we set this to the dave object if sayHello() is a function (i.e., not a method)? We can use apply can do just that that! Keep in mind that apply takes in two arguments: the object that this should refer to, and an array of arguments meant to be passed into the function (i.e. sayHello()). As such sayHello.apply(dave, ['Hello']); outputs the intended message: 'Hello, Dave. You're looking well today.'

Note that we also could have used call to solve this quiz. The expression sayHello.call(dave, 'Hello'); produces the same message: 'Hello, Dave. You're looking well today.'