Avoid mutation on js inheritance👴👦
Object.create() takes in a single object as an argument, and returns a new object. That new object's __proto__ property is set to whatever was originally passed into Object.create().

In this quiz, Parent.prototype was the argument passed into Object.create(). The return value of the expressionObject.create(Parent.prototype);was then set to the value of theChildconstructor'sprototypeproperty. After that, we instantiate a new object:child`.

The expression child instanceof Parent; returns a boolean indicating whether the Parent constructor exists in the child object's prototype chain. Since we know this is true after executing the first expression (i.e., Child.prototype = Object.create(Parent.prototype);), the console outputs true.